requirements.txt
Flask==2.2.5                 # Web framework
tensorflow==2.10.0           # TensorFlow (includes Keras API)
keras==2.10.0
numpy==1.23.5                # Numerical computing, compatible with TF 2.10
pandas==1.5.3                # Data handling
scikit-learn==1.1.3          # ML utilities (if used)
nltk==3.8.1                  # NLP preprocessing
gunicorn==20.1.0             # Production WSGI server for Flask
flask-mail==0.9.1            # Sending emails from Flask app
python-dotenv==1.0.0         # Load environment variables from .env files
